secrets package classes moved to domain and data package
