+++ app/src/main/java/com/ably/tracking/demo/publisher/di/KoinModule.kt
@@ -15,14 +15,14 @@ import com.ably.tracking.demo.publisher.data.api.buildDeliveryServiceApi
 import com.ably.tracking.demo.publisher.data.api.buildOkHttpClient
 import com.ably.tracking.demo.publisher.data.api.buildRetrofit
 import com.ably.tracking.demo.publisher.common.NotificationProvider
-import com.ably.tracking.demo.publisher.domain.DefaultOrderInteractor
-import com.ably.tracking.demo.publisher.domain.OrderManager
-import com.ably.tracking.demo.publisher.secrets.AndroidBase64Encoder
-import com.ably.tracking.demo.publisher.secrets.Base64Encoder
-import com.ably.tracking.demo.publisher.secrets.InMemorySecretsManager
-import com.ably.tracking.demo.publisher.secrets.SecretsManager
-import com.ably.tracking.demo.publisher.secrets.SecretsStorage
-import com.ably.tracking.demo.publisher.secrets.SharedPreferencesSecretsStorage
+import com.ably.tracking.demo.publisher.domain.order.DefaultOrderInteractor
+import com.ably.tracking.demo.publisher.domain.order.OrderManager
+import com.ably.tracking.demo.publisher.data.secrets.AndroidBase64Encoder
+import com.ably.tracking.demo.publisher.domain.secrets.Base64Encoder
+import com.ably.tracking.demo.publisher.domain.secrets.InMemorySecretsManager
+import com.ably.tracking.demo.publisher.domain.secrets.SecretsManager
+import com.ably.tracking.demo.publisher.domain.secrets.SecretsStorage
+import com.ably.tracking.demo.publisher.data.secrets.SharedPreferencesSecretsStorage
 import com.ably.tracking.demo.publisher.presentation.screens.login.LoginViewModel
 import com.ably.tracking.demo.publisher.presentation.screens.main.MainViewModel
 import com.ably.tracking.demo.publisher.presentation.navigation.AndroidJetpackNavigator
