+++ app/src/main/java/com/ably/tracking/demo/publisher/domain/order/OrderManager.kt
@@ -1,10 +1,11 @@
-package com.ably.tracking.demo.publisher.domain
+package com.ably.tracking.demo.publisher.domain.order
 
 import com.ably.tracking.TrackableState
 import com.ably.tracking.demo.publisher.data.ably.AssetTracker
 import com.ably.tracking.demo.publisher.common.copyAndReplaceElementAt
 import com.ably.tracking.demo.publisher.common.toOrderState
-import com.ably.tracking.demo.publisher.secrets.SecretsManager
+import com.ably.tracking.demo.publisher.domain.DeliveryServiceDataSource
+import com.ably.tracking.demo.publisher.domain.secrets.SecretsManager
 import com.ably.tracking.publisher.Trackable
 import kotlin.coroutines.CoroutineContext
 import kotlinx.coroutines.CoroutineDispatcher
