+++ app/src/main/java/com/ably/tracking/demo/publisher/domain/secrets/Base64Encoder.kt
@@ -1,4 +1,4 @@
-package com.ably.tracking.demo.publisher.secrets
+package com.ably.tracking.demo.publisher.domain.secrets
 
 interface Base64Encoder {
     fun encode(input: String): String
