+++ app/src/test/java/com/ably/tracking/demo/publisher/secrets/JavaBase64Encoder.kt
@@ -1,5 +1,6 @@
 package com.ably.tracking.demo.publisher.secrets
 
+import com.ably.tracking.demo.publisher.domain.secrets.Base64Encoder
 import java.util.Base64
 
 class JavaBase64Encoder : Base64Encoder {
